refactor: 투두 아이템 스타일 코드 정리
