+++ src/components/Member/MemberTodoItem.tsx
@@ -1,6 +1,6 @@
 import { Check } from '@/assets/icons';
 import { IconWrapper } from '@/components';
-import { checkedStyle, checkIconStyle, contentStyle, getStatusStyles, todoItemStyle } from '@/components/Todo';
+import { checkedStyle, checkIconStyle, contentWrapperStyle, getStatusStyles, todoItemStyle } from '@/components/Todo';
 import { TODO_STATUS } from '@/constants/todo';
 import { ToDoDetail } from '@/types';
 import { css, useTheme } from '@emotion/react';
@@ -10,7 +10,7 @@ const MemberTodoItem = ({ todo }: { todo: ToDoDetail }) => {
   const theme = useTheme();
   return (
     <li css={[todoItemStyle, getStatusStyles(isCompleted, theme), memberTodoItemStyle]}>
-      <div css={contentStyle}>
+      <div css={contentWrapperStyle}>
         <IconWrapper
           aria-label={isCompleted ? '완료' : '미완료'}
           aria-checked={isCompleted}
