+++ src/components/Todo/index.ts
@@ -1,4 +1,5 @@
 import TodoForm from '@/components/Todo/TodoForm';
 
+export * from './TodoItem';
 export * from './TodoList';
 export { TodoForm };
